Added configuration files, and flags to make video recording optional, but it's broken.

+++ wspsClient.sh
@@ -8,39 +8,49 @@
 # if the stream is not present.  So, if there's no video stream
 # it just won't create a video file
 
-streamPort=1234
-streamHost=169.254.193.113
-streamUser=tjobe
-audioDevice=default
-#localFolderBase="/AudioFiles"
-localFolderBase="${HOME}/Documents/AudioFiles"
-remoteFolder="${HOME}/Documents/AudioFiles/remote"
-itsoffset=0
-#webUser=terry
-webUser=tjobe
-#webHost=peopleforjesus.org
-webHost=localhost
-ffmpegExeLocal=~/win/bin/ffmpeg
-ffmpegExeRemote='C:\Program Files\ffmpeg\bin\ffmpeg.exe'
-basenameFfmpegExeRemote=`basename "${ffmpegExeRemote//\\\\//}"`
-
 # Parse command line options
-while getopts "y" opt; do
+while getopts "yvarl" opt; do
   case $opt in
   y)
-    forceOverwrite=1
+    y=1
+  ;;
+  v)
+    v=1
+  ;;
+  a)
+    v=0
+  ;;
+  r)
+    r=1
+  ;;
+  l)
+    r=0
   ;;
   \?)
-    Usage: $0 [-y]
+    Usage: $0 [-yva]
   ;;
-  esac 
+  esac
 done
+# defaults for flags
+forceOverwrite=${y-1}
+videoAndAudio=${v-1}
+remoteSync=${r-1}
 
+# Shift all processed options away
+shift $((OPTIND-1))
+configFile=${1-wspsDefaults.cfg}
+source "$configFile"
+
+# name that ffmpeg process will have on the windows side so you can kill it.
+basenameFfmpegExeRemote=`basename "${ffmpegExeRemote//\\\\//}"`
+
+# formatted date for file names
 dt=`date +%Y-%m-%d`
 
 # Get the local ip address
 # The Linux version assume that you're on the same subnet as the stream server
-os=`uname` localIP=""
+os=`uname`
+localIP=""
 case $os in
   Linux) localIP=`ifconfig | grep 'inet addr:' | grep ${streamHost:0:7}| cut -d: -f2 | awk '{print $1}'`;;
   Darwin|FreeBSD|OpenBSD) localIP=`ifconfig | grep -E 'inet.[0-9]' | grep -v '127.0.0.1' | awk '{ print $2}'` ;;
@@ -90,7 +100,7 @@ remoteVideoFile="${remoteFolder}/${videoFile}"
 remoteAudioFile="${remoteFolder}/${audioFile}"
 
 # Check for existence of files and ask if you want to overwrite
-if [[ ( ! -z ${forceOverwrite} ) && ( -e ${localVideoFile} || -e ${localAudioFile} || -z `ssh ${webUser}@${webHost} "ls ${remoteVideoFile} ${remoteAudioFile} 2>/dev/null"` ) ]]; then
+if [[ ( ! ${forceOverwrite} -eq 0 ) && ( -e ${localVideoFile} || -e ${localAudioFile} || -z `ssh ${webUser}@${webHost} "ls ${remoteVideoFile} ${remoteAudioFile} 2>/dev/null"` ) ]]; then
   read -p 'One or more of the files already exist, do you want to overwrite?(y/N):' a
   if [[ $a == [Nn] || $a == "" ]]; then
     echo Exiting.
@@ -102,11 +112,13 @@ fi
 function cleanup() {
   trap "" EXIT INT
   # back to front
-  [[ ! -z "$ffmpeg_video_pid" ]] && kill -9 "${ffmpeg_video_pid}"
+  if (( $videoAndAudio )); then
+    [[ ! -z "$ffmpeg_video_pid" ]] && kill -9 "${ffmpeg_video_pid}"
+    ssh ${streamUser}@${streamHost} "cmd /c taskkill /im \"${basenameFfmpegExeRemote}\" /f" 2>&1 >>${streamLog}
+    [[ ! -z "$stream_pid" ]] && kill -9 "${stream_pid}"
+  fi
   [[ ! -z "$lame_audio_pid" ]] && kill -9 "${lame_audio_pid}"
   [[ ! -z "$sox_pid" ]] && kill -9 "${sox_pid}"
-  ssh ${streamUser}@${streamHost} "cmd /c taskkill /im \"${basenameFfmpegExeRemote}\" /f" 2>&1 >>${streamLog}
-  [[ ! -z "$stream_pid" ]] && kill -9 "${stream_pid}"
 }
 
 trap "cleanup" EXIT INT
@@ -115,14 +127,12 @@ trap "cleanup" EXIT INT
 soxFifo=/tmp/sox.wav
 lameFifo=/tmp/lame.mp3
 ffmpegVideoFifo=/tmp/ffmpeg.mp4
-audioLocalFifo=/tmp/local.mp3
-videoLocalFifo=/tmp/local.mp4
 audioToVideoFifo=/tmp/audioToVideo.mp3
 audioToFileFifo=/tmp/audioToFile.mp3
 audioRemoteFifo=/tmp/remote.mp3
 videoRemoteFifo=/tmp/remote.mp4
 
-for fifo in "${soxFifo}" "${lameFifo}" "${ffmpegVideoFifo}" "${audioToVideoFifo}" "${audioToFileFifo}" "${audioLocalFifo}" "${videoLocalFifo}" "${audioRemoteFifo}" "${videoRemoteFifo}" ; do
+for fifo in "${soxFifo}" "${lameFifo}" "${ffmpegVideoFifo}" "${audioToVideoFifo}" "${audioToFileFifo}" "${audioRemoteFifo}" "${videoRemoteFifo}" ; do
   rm -f "$fifo"
   mkfifo "$fifo"
 done
@@ -132,14 +142,15 @@ streamLog=/tmp/stream.log
 soxLog=/tmp/sox.log
 lameLog=/tmp/lame.log
 ffmpegVideoLog=/tmp/ffmpegVideo.log
-audioLocalLog=/tmp/audioLocal.log
-videoLocalLog=/tmp/videoLocal.log
 audioRemoteLog=/tmp/audioRemote.log
 videoRemoteLog=/tmp/videoRemote.log
 
 # video streaming command
-ssh ${streamUser}@${streamHost} "cmd /c \"C:\\Program Files\\ffmpeg\\bin\\ffmpeg.exe\" -y -f dshow -t 3600 -i video=UScreenCapture -vcodec libx264 -vprofile high -preset slow -b:v 50k -maxrate 500k -bufsize 0k -vf scale=426:320 -pix_fmt yuv420p -metadata comment=trial -f mpegts udp://${localIP}:${streamPort}" 2>&1 >${streamLog} &
-stream_pid=$!
+# TODO: This path to ffmpeg is hard coded here, but set in a variable elsewhere
+if (( $videoAndAudio)); then
+  ssh ${streamUser}@${streamHost} "cmd /c \"C:\\Program Files\\ffmpeg\\bin\\ffmpeg.exe\" -y -f dshow -t 3600 -i video=UScreenCapture -vcodec libx264 -vprofile high -preset slow -b:v 50k -maxrate 500k -bufsize 0k -vf scale=426:320 -pix_fmt yuv420p -metadata comment=trial -f mpegts udp://${localIP}:${streamPort}" 2>&1 >${streamLog} &
+  stream_pid=$!
+fi
 
 # audio recording command
 sox -t "${audioDriver}" "${audioDevice}" -t wav "${soxFifo}" compand 0.2,0.20 5:-60,-40,-10 -5 -90 0.1 2>"${soxLog}" &
@@ -149,41 +160,54 @@ sox_pid=$!
 lame -m s -a -q 7 -V 6 "${soxFifo}" "${lameFifo}" 2>"${lameLog}" &
 lame_audio_pid=$!
 
-# audio to video piping command
-tee "${audioToFileFifo}" >"${audioToVideoFifo}" <"${lameFifo}" &
-audioToVideo_tee_pid=$!
+if (( $videoAndAudio )); then
+  # audio to video piping command
+  tee "${audioToFileFifo}" >"${audioToVideoFifo}" <"${lameFifo}" &
+  audioToVideo_tee_pid=$!
 
-# video recording command
-"${ffmpegExeLocal}" -loglevel verbose -f mpegts -i "udp://localhost:${streamPort}" -itsoffset ${itsoffset} -f mp3 -i "${audioToVideoFifo}" -y -c:v libx264 -f mp4 -movflags frag_keyframe+empty_moov "${ffmpegVideoFifo}" 2>${ffmpegVideoLog} &
-ffmpeg_video_pid=$!
+  # video recording command
+  "${ffmpegExeLocal}" -loglevel verbose -f mpegts -i "udp://localhost:${streamPort}" -itsoffset ${itsoffset} -f mp3 -i "${audioToVideoFifo}" -y -c:v libx264 -f mp4 -movflags frag_keyframe+empty_moov "${ffmpegVideoFifo}" 2>${ffmpegVideoLog} &
+  ffmpeg_video_pid=$!
+else
+  tee "${audioToFileFifo}" <"${lameFifo}"
+fi
 
-# audio piping command
-tee "${localAudioFile}" >"${audioRemoteFifo}" <"${audioToFileFifo}" &
-audio_tee_pid=$!
+if (( $remoteSync )); then
+  # audio piping command
+  tee "${localAudioFile}" >"${audioRemoteFifo}" <"${audioToFileFifo}" &
+  audio_tee_pid=$!
 
-# video piping command
-tee "${videoLocalFifo}" >"${videoRemoteFifo}" <"${ffmpegVideoFifo}" &
-video_tee_pid=$!
+  # audio ssh command
+  ssh ${webUser}@${webHost} "bash -c 'cat > \"${remoteAudioFile}\"'" <"${audioRemoteFifo}" &
+  audio_ssh_pid=$!
 
-# audio ssh command
-ssh ${webUser}@${webHost} "bash -c 'cat > \"${remoteAudioFile}\"'" <"${audioRemoteFifo}" &
-audio_ssh_pid=$!
+  if (( $videoAndAudio )); then
 
-# video ssh command
-ssh ${webUser}@${webHost} "bash -c 'cat > \"${remoteVideoFile}\"'" <"${videoRemoteFifo}" &
-video_ssh_pid=$!
+    # video piping command
+    tee "${localVideoFile}" >"${videoRemoteFifo}" <"${ffmpegVideoFifo}" &
+    video_tee_pid=$!
 
-# audio local command
-cat <"${audioLocalFifo}" >"${localAudioFile}" &
-audio_local_pid=$!
+    # video ssh command
+    ssh ${webUser}@${webHost} "bash -c 'cat > \"${remoteVideoFile}\"'" <"${videoRemoteFifo}" &
+    video_ssh_pid=$!
 
-# video local command
-cat <"${videoLocalFifo}" >"${localVideoFile}" &
-video_local_pid=$!
+  fi
+
+else
+
+  # audio piping command
+  cat "${localAudioFile}" <"${audioToFileFifo}" &
+  audio_tee_pid=$!
+
+  if (( $remoteSync )); then
+    cat "${localVideoFile}" <"${ffmpegVideoFifo}" 
+  fi
+
+fi
 
 # wait for kill signal to stop recording
 wait ${stream_pid} ${sox_pid} ${lame_audio_pid} ${ffmpeg_video_pid} ${audioToVideo_tee_pid} ${audio_tee_pid} ${video_tee_pid}\
-${audio_ssh_pid} ${video_ssh_pid} ${audio_local_pid} ${video_local_pid}
+${audio_ssh_pid} ${video_ssh_pid}
 
 # remove the named pipes
 for ff in "${soxFifo}" "${lameFifo}" "${ffmpegVideoFifo}" "${audioToVideoFifo}" "${audioToFileFifo}" "${audioLocalFifo}" "${videoLocalFifo}" "${audioRemoteFifo}" "${videoRemoteFifo}" ; do
